Add files via upload
清理

+++ main.py
@@ -1,10 +1,13 @@
 import pandas as pd
 import json
 import time
-from 爬虫 import get_save,get_url,down_pic,down_pic_usual,fix_name,find_bangumi
+from 爬虫 import get_save, get_url, down_pic,down_pic_usual,fix_name
 from create_ico import reset
 import os
 import shutil
+import difflib
+import time
+from pyiter import  it
 
 #等以后把电影做进来了把电影单独划出去
 #不知道bangumi能不能搜小说和音乐
@@ -28,10 +31,13 @@ cartoon_rank1=[]#按观看顺序来排名
 cartoon_rank2=[]#按排名来排序,合并了同一个的
 cartoon_rank3=[]
 #init_menu="C:/Users/17402/Desktop/爬虫/"#输出不同目录 可更改部分
-init_menu="C:/Users/17402/Desktop/爬虫/new/"
-#total_menu="C:/Users/17402/Desktop/爬虫/"
-total_menu="C:/Users/17402/Desktop/爬虫/total/"
-excel_url="C:/Users/17402/Desktop/爬虫/new/番剧记录平野.csv"
+#init_menu="C:/Users/17402/Desktop/爬虫/new/"
+#total_menu="C:/Users/17402/Desktop/爬虫/total/"
+#excel_url="C:/Users/17402/Desktop/爬虫/new/番剧记录平野.csv"
+
+init_menu="C:/Users/17402/Desktop/爬虫/new1/"#输出目录
+total_menu="C:/Users/17402/Desktop/爬虫/total/"#总数据存储目录
+excel_url="C:/Users/17402/Desktop/爬虫/new1/1.csv"#看番excel存储地址
 def open_excel(excel_url):#打开excel
     global cartoon_rank1,cartoon_rank2,cartoon_rank3
     cartoon_rank1=[]
@@ -56,6 +62,7 @@ def open_excel(excel_url):#打开excel
         if b.排名.item()!=0:
             try:
                 if cartoon_list.get(int(b.bangumi.item()))!=None:
+                    print(cartoon_list[int(b.bangumi.item())])
                     print(b.名字.item())
                     return False
                 cartoon_list[int(b.bangumi.item())]=[int(b.排名.item()),int(b.季数.item()),b.名字.item(),b.注释.item(),int(b.index.item())]
@@ -87,23 +94,33 @@ def isexist_data():#补充数据库内容
             content1 = get_url('https://bangumi.tv/subject/' + x + '/characters')
             save_dict = get_save(content, content1, x,person_data_total,character_data_total)
             cartoon_data_total[x] = save_dict
-            if not os.path.exists(init_menu + "save/total_img/" + x + '.jpg'):
+            if not os.path.exists(total_menu + "save/total_img/" + x + '.jpg'):
                 time.sleep(1)
-                down_pic(content, init_menu + "save/total_img/" + x + '.jpg')
+                down_pic(content, total_menu + "save/total_img/" + x + '.jpg')
             time.sleep(2)
 
 
 def isexist_data2(): #补充肖像类文件
-    for x in person_data.keys():
+    now=time.time()
+    count0=len(list(x for x in person_data_total.keys() if person_data_total[x]['download']==False and person_data_total[x]['img']!=None))
+    count1=len(list(x for x in character_data_total.keys() if character_data_total[x]['download']==False and character_data_total[x]['img']!=None))
+    count3=0
+    for x in person_data_total.keys():
         if person_data_total[x]['download']==False and person_data_total[x]['img']!=None:
+            count3+=1
+            if count3%100==0:
+                print(str((time.time()-now)*((count1+count0-count3)/count3)//60)+'分钟')
             if down_pic_usual(person_data_total[x]['img'],total_menu+'save/total_img/person/'+str(x)+'.jpg')==False:
                 person_data_total[x]['img']=None
                 continue
             person_data_total[x]['download'] = True
             #time.sleep(0.5)
     print('perpson载入完成')
-    for x in character_data.keys():
+    for x in character_data_total.keys():
         if character_data_total[x]['download']==False and character_data_total[x]['img']!=None:
+            count3+=1
+            if count3%100==0:
+                print(str((time.time()-now)*((count1+count0-count3)/count3)//60)+'分钟')
             if down_pic_usual(character_data_total[x]['img'],total_menu+'save/total_img/character/'+str(x)+'.jpg')==False:
                 character_data_total[x]['img']=None
                 continue
@@ -115,15 +132,18 @@ def copy():
         x=str(x)
         for dic in cartoon_data_total[x]['cv']:
             person_data[dic['p_id']]=person_data_total[dic['p_id']]
-            if person_data_total[dic['p_id']]['img'] != None:
+            if person_data_total[dic['p_id']]['img'] != None :
                 src = total_menu + 'save/total_img/person/' + str(dic['p_id']) + '.jpg'
                 tar = init_menu + 'save/total_img/person/' + str(dic['p_id']) + '.jpg'
-                shutil.copy(src, tar)
-            character_data[dic['c_id']] = character_data_total[dic['c_id']]
-            if character_data_total[dic['c_id']]['img'] != None:
-                src = total_menu + 'save/total_img/character/' + str(dic['c_id']) + '.jpg'
-                tar = init_menu + 'save/total_img/character/' + str(dic['c_id']) + '.jpg'
-                shutil.copy(src, tar)
+                if not os.path.exists(tar):
+                    shutil.copy(src, tar)
+            for ch1 in person_data[dic['p_id']]['ch']:#统计声优关联角色的时候要给出数据库里全部的角色
+                character_data[ch1] = character_data_total[ch1]
+                if character_data_total[dic['c_id']]['img'] != None :
+                    src = total_menu + 'save/total_img/character/' + str(ch1) + '.jpg'
+                    tar = init_menu + 'save/total_img/character/' + str(ch1) + '.jpg'
+                    if not os.path.exists(tar):
+                        shutil.copy(src, tar)
 
     for x in cartoon_list.keys():
         x=str(x)
@@ -139,7 +159,7 @@ def check_isexist_update_ch_and_pe():
                 rl = total_menu + 'save/total_img/'+url+'/' + str(x) + '.jpg'
                 if os.path.exists(rl):
                     if os.stat(rl).st_size <= 512:
-                        print(x)
+                        print(url+"重下载"+x)
                         try:
                             if down_pic_usual(person_data[x]['img'],rl) !=True:
                                 person_data[x]['download'] = False
@@ -151,6 +171,21 @@ def check_isexist_update_ch_and_pe():
 
     Multiplexing(person_data_total,'person')
     Multiplexing(character_data_total, 'character')
+    for x in cartoon_data_total.keys():
+            rl = total_menu + 'save/total_img/' + str(x) + '.jpg'
+            if os.path.exists(rl):
+                if os.stat(rl).st_size <= 512:
+                    print('封面重下载'+x)
+                    try:
+                        content = get_url('https://bangumi.tv/subject/' + x)
+                        time.sleep(1)
+                        down_pic(content, total_menu + "save/total_img/" + x + '.jpg')
+                    except:
+                        print(x+'fail')
+            else:
+                content = get_url('https://bangumi.tv/subject/' + x)
+                time.sleep(1)
+                down_pic(content, total_menu + "save/total_img/" + x + '.jpg')
 def check_isexist_update_ch_and_pe2():#补一下设定 文件大于大小则把设定补充为true
     def Multiplexing(person_data,url):
         for x in person_data.keys():
@@ -166,7 +201,7 @@ def isexist_content():#填充目录内容
     len_al=len(alist)
     for x in range(len(cartoon_rank2)):
         name=cartoon_data[str(cartoon_rank2[x][0])]['中文名']
-        name_rl=str(x+1)+"."+name
+        name_rl=(str(x+1)+"."+name).replace('/','').replace('!','').replace("?",'')
         if name_rl not in alist:#判断原有在不在，不在就开删然后建
             os.mkdir(con+name_rl)
             if x<len_al:#删除有变更的内容
@@ -197,11 +232,16 @@ def data_strip():
                     cartoon_data_total[x][y]=cartoon_data_total[x][y].strip()
 def cover_none():
     for x in cartoon_data_total.keys():
-        if '播放结束' not in x:
+        if '放送开始' not in cartoon_data_total[x].keys():
+            cartoon_data_total[x]['放送开始']=None
+            cartoon_data_total[x]['播放结束'] = None
+            continue
+        if '播放结束' not in cartoon_data_total[x].keys():
             add_time=time.mktime(time.strptime(cartoon_data_total[x]['放送开始'],"%Y年%m月%d日"))+60*60*24*7*int(cartoon_data_total[x]['话数'])
             cartoon_data_total[x]['播放结束']=time.strftime("%Y年%m月%d日",time.localtime(add_time))
             #print(cartoon_data[x]['放送开始'])
             #print(cartoon_data[x]['播放结束'])
+
 def add_person(): #之前写代码不完善用的补充技巧 实际上也不完善，这个函数根本用不了
     for x in person_data:
         person_data[x]['ch'] = 0
@@ -239,6 +279,14 @@ def isexist_index():#没有目录就创建
         os.mkdir(init_menu+"save/overleaf/images/person")
     if "character" not in os.listdir(init_menu+"save/overleaf/images"):
         os.mkdir(init_menu+"save/overleaf/images/character")
+    if "test" not in os.listdir(init_menu+"save/"):
+        os.mkdir(init_menu+"save/test")
+    if "images" not in os.listdir(init_menu+"save/test"):
+        os.mkdir(init_menu+"save/test/images")
+    if "person" not in os.listdir(init_menu+"save/test/images"):
+        os.mkdir(init_menu+"save/test/images/person")
+    if "character" not in os.listdir(init_menu+"save/test/images"):
+        os.mkdir(init_menu+"save/test/images/character")
 
 def upload_json_init():
     global cartoon_rank1,cartoon_rank2,cartoon_data,person_data,character_data
@@ -254,7 +302,7 @@ def upload_json_init():
             total_data=save_json
             cartoon_data,person_data,character_data=total_data
         except Exception as e:
-            print("传入失败"+e)
+            print("传入失败"+str(e))
 
 def upload_json():
     global cartoon_data_total,person_data_total,character_data_total
@@ -264,7 +312,7 @@ def upload_json():
             total_data_total=save_json
             cartoon_data_total,person_data_total,character_data_total=total_data_total
         except Exception as e:
-            print("源数据传入失败"+e)
+            print("源数据传入失败"+str(e))
 
 def load_json_init():#将已经有的数据存进去
     with open(init_menu+"save/struct.json","w") as m:
@@ -298,25 +346,127 @@ def read_e(pri=False):#读入excel并进行初始化
     fix_name(cartoon_data_total) #看看有没有不存在的中文名，没有则补充
     print('修补完成')
     load_json() #存入源数据
+    check_isexist_update_ch_and_pe2()
     isexist_data2()
-    #check_isexist_update_ch_and_pe2()
-    #check_isexist_update_ch_and_pe()
+    load_json()
+    check_isexist_update_ch_and_pe2()
+    check_isexist_update_ch_and_pe()
     load_json()
     copy()#复制一份数据
     load_json_init()
     print('开始创建目录')
     isexist_content()
-'''upload_json()
-for keys in cartoon_data_total.keys():
-    m=cartoon_data_total[keys]['cv']
-    cartoon_data_total[keys]['cv']=list(it(m).map(lambda x: {"p_id":x[3],"c_id":x[2]}))
-load_json()'''
+
+def get_newest_information():
+   m=get_url('https://unpkg.com/bangumi-data@0.3/dist/data.json')
+   with open('data.json','w',encoding='utf-8') as w:
+       w.write(m)
+
+def get_match():
+    big = []
+    def big_bing(a):
+        for x in a:
+            if type(x) == list:
+                big_bing(x)
+            else:
+                big.append(x)
+
+    def get_bangumi():
+        with open('data.json', encoding='utf-8') as w:
+            a = json.load(w)  # sitemate items:[{}]
+            b = []
+            for x in a['items'][0].keys():
+                print(x, a['items'][0][x])
+            for x in a['items']:
+                if x['type'] == 'tv':
+                    b.append([list(x['titleTranslate'][m] for m in x['titleTranslate'].keys()),
+                              list(m['id'] for m in x['sites'] if m['site'] == 'bangumi')])
+            b = list(x for x in b if len(x[1]) == 1)
+        return b
+    a=get_bangumi()
+    search_data=list(x[0] for x in a)
+    big_bing(search_data)
+    search_dict=dict()
+    for x in a:
+        for y in x[0]:
+            for z in y:
+                search_dict[z]=x[1][0]
+    open_excel(excel_url)
+    upload_json()
+    m=[]
+    n=0
+    for x in cartoon_rank3:
+        #dic={'0':'','1':'一','2':'二','3':'三','4':'四','5':'五','6':'六','7':'七'}str(dic[str(x[1])])
+        result=difflib.get_close_matches(x[2],search_dict.keys(),10)
+        for y in result:
+            if search_dict[y] in m :
+                pass
+            else:
+                m.append(search_dict[y])
+                if int(search_dict[y]) !=x[3]:
+                    print(str(search_dict[y]),str(x[3]))
+                    print(x[2],y)
+                    n+=1
+                break
+    print(len(cartoon_rank3))
+    print(n,n/len(cartoon_rank3))
+def get_match1(in_e,out_e):
+    big = []
+
+    def big_bing(a):
+        for x in a:
+            if type(x) == list:
+                big_bing(x)
+            else:
+                big.append(x)
+
+    def get_bangumi():
+        with open('data.json', encoding='utf-8') as w:
+            a = json.load(w)  # sitemate items:[{}]
+            b = []
+            for x in a['items'][0].keys():
+                print(x, a['items'][0][x])
+            for x in a['items']:
+                    b.append([list(x['titleTranslate'][m] for m in x['titleTranslate'].keys()),
+                              list(m['id'] for m in x['sites'] if m['site'] == 'bangumi')])
+            b = list(x for x in b if len(x[1]) == 1)
+        return b
+
+    a = get_bangumi()
+    search_data = list(x[0] for x in a)
+    big_bing(search_data)
+    search_dict = dict()
+    for x in a:
+        for y in x[0]:
+            for z in y:
+                search_dict[z] = x[1][0]
+    with open(in_e,'r',encoding='utf-8') as m:
+        m=list(x.strip() for x in m.readlines())
+        print(m)
+    print()
+    with open(out_e,'w',encoding='utf-8') as w:
+        m1=[]
+        for x in m:
+            result = difflib.get_close_matches(x, search_dict.keys(),10,0.6)
+            for y in result:
+                if search_dict[y] in m1:
+                    pass
+                else:
+                    m1.append(search_dict[y])
+                    break
+            else:
+                print(x)
+                continue
+            if len(result)==0:
+                print(x)
+            w.writelines(x+','+search_dict[y]+'\n' if len(result) > 0 else '')
+#get_match1("C:/Users/17402/Desktop/爬虫/new1/番剧记录 - 副本.csv","C:/Users/17402/Desktop/爬虫/new1/0.csv")
 #read_e(True)
-#print(cartoon_data_total['316247'])
-#cartoon_data_total1=dict()#把不同分目录的数据都存在一个地方。明智的选择
-#person_data_total1=dict()
-#character_data_total1=dict()
-#with open('C:/Users/17402/Desktop/爬虫/' + "save/data.json", "w") as m:
-#    total_data_total1 = [cartoon_data_total1, person_data_total1, character_data_total1]
-#    json.dump(total_data_total1, m)
-#print(cartoon_data_total1['316247'])
+open_excel(excel_url)
+upload_json_init()
+upload_json()
+#copy()
+#load_json_init()
+#read_e(True)
+
+
